Implementación services base

+++ src/main/java/pe/datasys/multiisp/repo/ILiquidacionRepo.java
@@ -2,5 +2,5 @@ package pe.datasys.multiisp.repo;
 
 import pe.datasys.multiisp.model.Liquidacion;
 
-public interface ILiquidacionRepo extends IGenericRepo<Liquidacion, Long> {
+public interface ILiquidacionRepo extends IGenericRepo<Liquidacion, Integer> {
 }
